Point integration tests to enpa-integration-testing Firebase Project (#15)
Change integration tests so that they point to a true Firestore instance
because the Firestore Emulator has not implemented handling for
the PartitionQuery method.

Change test to enable FirestoreClient to connect to the FirestoreEmulator.
Update the FirestoreClient in the FirestoreConnector as well so that it connects to the emulator when the pipeline is being run in a test.