Clean up project usage (#99)
* Change startTime defaulting logic for deletion

When startTime is not supplied, make startTime = now - deleteBackwardsWindow. In this scenario, all documents between startTime - 2 * deleteBackwardsWindow -> startTime - deleteBackwardsWindow will be deleted. DeleteBackwardsWindow defaults to 24 hours, so all documents uploaded between 48 and 24 hours before runtime will be deleted if both startTime and deleteBackwardsWindow are not specified.

Use DataflowPipelineOptions::getProject rather than setting FirebaseProjectId in pipeline options since we are now using the convention of keeping the Firestore instance and Dataflow job that accesses the instance in the same project.

Co-authored-by: Stephanie Wang <stephaniewang526@users.noreply.github.com>