fix new deterministic batch id scheme (#102)
* fix new deterministic batch id scheme

When shares were output in multiple batches, the batch id
would resolve to the same uuid and we'd overwrite the same
file. Although the batch uuid was being seeded with metric
name and date, it also needs to include some differentiation
for which batch part it is. This just numbers the batches at
the same point where we create the groupings. This won't
guarantee that the same data shares end up in the same files
(we don't actually need that) but it does guarantee multiple
runs over same window generate exactly the same files.