remove profiles (#104)
Now that we're extending from DataflowPipelineOptions we pull in
the dataflow runner jars all the time anyways.

Also, restore ability to not specify a manifest (and manually pass
in pha/facilitator output locations), and remove synchronization
that's no longer necessary now that init() is happening just once
in static main context.