switch from numbering batches to uuids (#194)
In order to assign deterministic numbers we make use of shared
state, which for very small values of batch size, leads to a lot
of contention. Switch to assigning a random uuid.

This foregoes the nice property that a rerun of the pipeline will
result in same files being written at downstream partner storage.
Instead this just preserves the property that when an individual
batch write is retried (by the beam runner) it won't create
duplicates.

This leads to a performance boost for small values of batch size,
though in practice we run with values of O(100k). E.g.
Compare batch size of 1, ~10K shares, 70% reduction in runtime:
https://console.cloud.google.com/dataflow/jobs/us-west2/2021-01-21_17_39_52-9465626085867489723?project=enpa-ingestion-ta-ta
https://console.cloud.google.com/dataflow/jobs/us-central1/2021-01-27_18_31_07-5544186918352481055?project=enpa-load-test