extract a superclass attestations (#196)
This creates a package for attestations, and establishes the
basic interface the pipeline will use. It also pulls up the
check attestation flag earlier into the pipeline (so no
attestation logic will execute when off), and carves out
options specific to key attestation.

This can be the first step to adding new attestation types

Tested unit tests and integration tests as well as running the
pipeline on load test set with attestation off as well as on
(with key attestation specific flags set).