refactoring (#198)
This switches from one maven project to a multi-module project with the
top project being an aggregator. I took the approach of both inheritance
and aggregation so that we could define some common properties for all
the modules but also spell out the relationships and dependencies more
clearly.
http://maven.apache.org/guides/introduction/introduction-to-the-pom.html#Project_Inheritance_vs_Project_Aggregation

Since there aren't a lot of java files involved here, I thought it made
sense to still use a single source directory (rather than have deep
subdirectories with a single java file):
http://maven.apache.org/guides/introduction/introduction-to-the-pom.html#Project_Inheritance_vs_Project_Aggregation

This took some finagling to get the test sources, annotation processors,
and code generators all wired up correctly, but I think its a bit
clearer now as to what depends on what, rather than the one big ball of
spaghetti it was before.

Co-authored-by: Robert Kubis <hostirosti@users.noreply.github.com>